+++ graphene_django_crud/settings.py
@@ -4,7 +4,7 @@ from django.utils.module_loading import import_string
 
 DEFAULTS = {
     "DEFAULT_CONNECTION_NODES_FIELD_NAME": "data",
-    "FILE_TYPE_ACTIVE_CONTENT_FIELD": False,
+    "FILE_TYPE_CONTENT_FIELD_ACTIVE": False,
 }
 
 IMPORT_STRINGS = ()