+++ graphene_django_crud/__init__.py
@@ -3,7 +3,7 @@ from .types import DjangoGrapheneCRUD, DjangoCRUDObjectType, resolver_hints
 from .utils import where_input_to_Q, order_by_input_to_args
 from .base_types import DefaultConnection, EmptyDefaultConnection
 
-__version__ = "1.3.0"
+__version__ = "1.3.1"
 
 __all__ = (
     "__version__",